+++ cmd/api/main.go
@@ -7,10 +7,14 @@ import (
 	"time"
 
 	defaultresolver "github.com/Chatterino/api/internal/resolvers/default"
+	"github.com/Chatterino/api/internal/routes/twitchemotes"
+	"github.com/Chatterino/api/internal/twitchapiclient"
+	"github.com/Chatterino/api/pkg/cache"
 	"github.com/Chatterino/api/pkg/config"
 	"github.com/Chatterino/api/pkg/resolver"
 	"github.com/Chatterino/api/pkg/thumbnail"
 	"github.com/go-chi/chi/v5"
+	"github.com/go-chi/chi/v5/middleware"
 )
 
 var (
@@ -65,9 +69,19 @@ func main() {
 
 	router := chi.NewRouter()
 
-	handleTwitchEmotes(router)
+	// Strip trailing slashes from API requests
+	router.Use(middleware.StripSlashes)
+
+	var helixUsernameCache *cache.Cache
+
+	helixClient, helixUsernameCache, err := twitchapiclient.New(cfg)
+	if err != nil {
+		log.Printf("[Twitch] %s\n", err.Error())
+	}
+
+	twitchemotes.Initialize(router, helixClient, helixUsernameCache)
 	handleHealth(router)
-	defaultresolver.Initialize(router, cfg)
+	defaultresolver.Initialize(router, cfg, helixClient)
 
 	listen(cfg.BindAddress, mountRouter(router))
 }