+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added link preview support for 7tv emote links. (#155)
 - Skip lilliput if image is below maxThumbnailSize. (#184)
+- Dev: Change Emote Set backend from `twitchemotes.com` to the Twitch Helix API. (#175)
 
 ## 1.2.0